+++ src/main/java/net/frozenblock/wilderwild/particle/PollenParticle.java
@@ -49,10 +49,10 @@ public class PollenParticle extends TextureSheetParticle {
 		this.setSize(0.01F, 0.02F);
 		this.pickSprite(spriteProvider);
 		this.quadSize *= this.random.nextFloat() * 0.6F + 0.6F;
-		this.lifetime = (int) (16.0D / (AdvancedMath.random().nextDouble() * 0.8D + 0.2D));
+		this.lifetime = (int) (16D / (AdvancedMath.random().nextDouble() * 0.8D + 0.2D));
 		this.hasPhysics = true;
-		this.friction = 1.0F;
-		this.gravity = 0.0F;
+		this.friction = 1F;
+		this.gravity = 0F;
 	}
 
 	@Override