+++ CHANGELOG.md
@@ -103,6 +103,7 @@ Bug Fixes & Changes
   - Prevented the Warden from moving while it is digging or emerging.
   - Changed the Warden's secret death sound from stereo to mono.
   - Refactored mobs' `canSpawn` methods to reflect Vanilla's, following the template `checkXSpawnRules.`
+  - Significantly improved Wilder Wild's biome parameter modifications and cut down `OverworldBiomeBuilderMixin` significantly in the process.
   - Pumpkins now generate much more frequently in Old Growth Dark Forests.
   - Changed the `Cherry Grove Placement` config to be disabled by default as it wasn't as immersive as anticipated.
   - Fixed the `Stony Shore Placement` config lang reading as `CStony Shore Placement.`